Factor out constants and selectors
